单元测试已覆盖100%

+++ tests/base_redis_test_case.py
@@ -2,16 +2,17 @@
 import unittest
 import redis
 from a3redis.bases import RedisClientFactory, RedisMode
-
+from a3redis.bases.redis_client_factory import DEFAULT_NAME
 
 ALL_CONF = {
+    # 还有一个默认的redis://127.0.0.1:6379/0，没有明写
     RedisMode.Standalone: {
         'mode': RedisMode.Standalone,
         'init': {
-            'host': 'redis-node-0.redis-headless',
+            'host': 'standalone-redis-master',
             'port': 6379,
             'db': 0,
-            'password': "123456",
+            'password': RedisMode.Standalone,
             'decode_responses': True
         }
     },
@@ -19,14 +20,14 @@ ALL_CONF = {
         'mode': RedisMode.Sentinel,
         'init': {
             'sentinels': [
-                ('redis-node-0.redis-headless', 26379),
-                ('redis-node-1.redis-headless', 26379),
-                ('redis-node-2.redis-headless', 26379)
+                ('sentinel-redis-node-0.sentinel-redis-headless', 26379),
+                ('sentinel-redis-node-1.sentinel-redis-headless', 26379),
+                ('sentinel-redis-node-2.sentinel-redis-headless', 26379)
             ],
             'sentinel_kwargs': {
-                'password': '123456'
+                'password': RedisMode.Sentinel
             },
-            'password': '123456',
+            'password': RedisMode.Sentinel,
             'db': 1,
             'decode_responses': True,
         },
@@ -38,8 +39,9 @@ ALL_CONF = {
         'mode': RedisMode.Cluster,
         'init': {
             'host': "cluster-redis-cluster",
-            'port': "6379",
-            'password': "123456",
+            'port': 6379,
+            'password': RedisMode.Cluster,
+            'decode_responses': True
         }
     }
 
@@ -48,8 +50,9 @@ ALL_CONF = {
 
 class BaseRedisTestCase(unittest.TestCase):
     rdb: redis.Redis = None
-    redis_name = 'default'
+    redis_name = DEFAULT_NAME
 
+    @classmethod
     def setUpClass(cls):
         RedisClientFactory.init_redis_clients(ALL_CONF)
 
@@ -57,4 +60,5 @@ class BaseRedisTestCase(unittest.TestCase):
         self.rdb = RedisClientFactory.get_rdb(self.redis_name)
 
     def tearDown(self) -> None:
-        self.rdb.flushdb()
+        if self.redis_name != RedisMode.Cluster:
+            self.rdb.flushdb()
